--- README.md
@@ -1,6 +1,0 @@
-# Melhorias
-- [ ] Bloquear mais de 6 cores.
-- [ ] Colocar input para alterar cor.
-- [ ] Salvar lista de cores gerada.
-- [ ] Editar um cor única da lista salva.
-- [ ] Opção para gerar somente cores claras.
