+++ src/components/HeaderComponent.js
@@ -1,0 +1,25 @@
+import { Header } from "../styles/Header";
+
+import { Input } from "../styles/Input";
+
+import { DropdownComponent } from "../components/DropdownComponent";
+import { UserComponent } from "../components/UserComponent";
+
+import logo from "../assets/images/logo.svg";
+
+const HeaderComponent = () => {
+  return (
+    <Header>
+      <div className="logo" >
+        <img src={logo} alt="The Colors" />
+        <strong>The Colors</strong>
+      </div>
+      <div id="bloco_usuario">
+        <UserComponent />
+        <DropdownComponent />
+      </div>
+    </Header>
+  );
+};
+
+export { HeaderComponent };
